+++ requirements.txt
@@ -1,0 +1,1 @@
+# Зависимости проекта
